Removed layout/VisualiserDashboardLayout, moved code to views/VisualiseDashboard
