+++ client/src/layout/index.ts
@@ -1,2 +1,1 @@
 export { default as HomepageLayout } from './HomepageLayout';
-export { default as VisualiserDashboardLayout } from './VisualiserDashboardLayout';
